created tests module, moved old tests.py to tests/model_tests.py, moved EtherpadLiteClientMock and testserver into tests/utils.py, started to write tests for the views (tests/view_tests.py)
