+++ etherpadlite/tests/__init__.py
@@ -1,0 +1,2 @@
+from model_tests import *
+from view_tests import *
