+++ etherpadlite/tests/model_tests.py
@@ -1,72 +1,18 @@
-"""
-This file demonstrates writing tests using the unittest module. These will pass
-when you run "manage.py test".
-
-Replace this with more appropriate tests for your application.
-"""
-
-from django.utils import unittest
+# ~ encoding: utf-8 ~
+from django.test import TestCase
 from django.contrib.auth.models import *
 
+from etherpadlite.tests.utils import *
 from etherpadlite.models import *
 import etherpadlite.models
 
+
 import string
-
-testserver = {
-    'title': 'Testserver',
-    'url': 'http://www.example.com',
-    'apikey': '1234567'
-}
-
-
-class EtherpadLiteClientMock(object):
-    """ A class to mock the EtherpadLiteClient for easy testing without having
-    a Etherpad-lite server running in the background.
-    """
-
-    def __init__(self, apiKey=None, baseUrl=None):
-        if apiKey:
-            self.apiKey = apiKey
-
-        if baseUrl:
-            self.baseUrl = baseUrl
-
-    def createGroupIfNotExistsFor(self, group_information):
-        """ Simply return the group_information 'as it is' instead of building
-        a cryptic hash.
-        """
-        return {'groupID': group_information}
-
-    def deleteGroup(self, *args):
-        pass
-
-    def createAuthorIfNotExistsFor(self, user_id, name=''):
-        """ Simply return the user_id 'as it is' instead of building a cryptic
-        hash.
-        """
-        return {'authorID': user_id}
-
-    def createGroupPad(self, group_id, name):
-        pass
-
-    def deletePad(self, pad_id):
-        pass
-
-    def getPublicStatus(self, pad_id):
-        pad_name = pad_id.split('$')[-1]
-        if pad_name == 'R2D2':
-            return {'publicStatus': True}
-        else:
-            return {'publicStatus': False}
-
-    def getReadOnlyID(self, pad_id):
-        return pad_id
 
 etherpadlite.models.EtherpadLiteClient = EtherpadLiteClientMock
 
 
-class PadServerTestCase(unittest.TestCase):
+class PadServerTestCase(TestCase):
     """Test cases for the Server model
     """
 
@@ -96,7 +42,7 @@ class PadServerTestCase(unittest.TestCase):
         self.server.delete()
 
 
-class PadGroupTestCase(unittest.TestCase):
+class PadGroupTestCase(TestCase):
     """Test cases for the Group model
     """
 
@@ -155,7 +101,7 @@ class PadGroupTestCase(unittest.TestCase):
         self.server.delete()
 
 
-class PadAuthorTestCase(unittest.TestCase):
+class PadAuthorTestCase(TestCase):
     """Test cases for the Author model
     """
 
@@ -189,7 +135,7 @@ class PadAuthorTestCase(unittest.TestCase):
         self.server.delete()
 
 
-class PadTestCase(unittest.TestCase):
+class PadTestCase(TestCase):
     """Test cases for the Pad model
     """
